# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tests Platform API",
    "description": "Assembly + administration engine for a large testing program.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/blueprints": {
      "post": {
        "tags": [
          "blueprints"
        ],
        "summary": "Create Blueprint",
        "operationId": "create_blueprint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Blueprint"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blueprints/generate": {
      "post": {
        "tags": [
          "blueprints"
        ],
        "summary": "Generate Blueprint From Curriculum",
        "description": "Curriculum\u2192blueprint generator (BP-MODES-1 \u00a76, rev. 2026-07-09).\n\nConsumes a curriculum manifest (inline, or by ``course_id`` from the catalog)\nand emits a blueprint for the requested \u00a76.2 test type (unit_quiz /\nmid_course / end_of_course / cumulative_final) using \u00a76.1 dimension-sum\nweights (median imputation, reported), largest-remainder rounding, authored\ncognitive profile, and per-binding TIF/cell rules. When ``pool_id`` is given\nthe blueprint is validated against that pool's tag counts (the \u00a76 gate). The\nblueprint is returned for review, not stored \u2014 persist via ``POST /blueprints``.",
        "operationId": "generate_blueprint_from_curriculum",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBlueprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateBlueprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blueprints/{blueprint_id}": {
      "get": {
        "tags": [
          "blueprints"
        ],
        "summary": "Get Blueprint",
        "operationId": "get_blueprint",
        "parameters": [
          {
            "name": "blueprint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Blueprint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Blueprint": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 2
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "untitled-blueprint"
          },
          "length": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Length",
            "description": "items per form"
          },
          "num_forms": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Num Forms",
            "default": 1
          },
          "content_constraints": {
            "items": {
              "$ref": "#/components/schemas/ContentConstraint"
            },
            "type": "array",
            "title": "Content Constraints"
          },
          "statistical_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TIFTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "enemy_policy": {
            "$ref": "#/components/schemas/EnemyPolicy"
          },
          "exposure_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExposureTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "exposure_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExposureFeedback"
              },
              {
                "type": "null"
              }
            ]
          },
          "segments": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          }
        },
        "type": "object",
        "required": [
          "length"
        ],
        "title": "Blueprint",
        "description": "Full assembly specification for one (or several parallel) forms.\n\nDelivery-mode-agnostic (BP-MODES-1): one blueprint is satisfiable by fixed-form,\nLOFT, and CAT. ``statistical_target`` is **optional** \u2014 a blueprint with none is a\n*content-only* blueprint (fixed-form then assembles for feasibility only, still\nreporting realized TIF; see ``docs/blueprint-delivery-mode-semantics.md`` \u00a72.1)."
      },
      "BlueprintRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "num_forms": {
            "type": "integer",
            "title": "Num Forms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "blueprint": {
            "$ref": "#/components/schemas/Blueprint"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "length",
          "num_forms",
          "created_at",
          "blueprint"
        ],
        "title": "BlueprintRead"
      },
      "CognitiveProfile": {
        "properties": {
          "dimension": {
            "type": "string",
            "enum": [
              "bloom_process",
              "bloom_knowledge",
              "timss"
            ],
            "title": "Dimension"
          },
          "distribution": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distribution"
          },
          "per_unit_minimums": {
            "items": {
              "$ref": "#/components/schemas/PerUnitCognitiveMinimum"
            },
            "type": "array",
            "title": "Per Unit Minimums"
          }
        },
        "type": "object",
        "required": [
          "dimension"
        ],
        "title": "CognitiveProfile",
        "description": "Authored cognitive requirements (never derived from the curriculum).\n\n``distribution`` (value \u2192 share of the form, shares summing to 1) is emitted as\nmarginal proportion constraints; ``per_unit_minimums`` as cross-classified\n{unit \u00d7 dimension} count minimums. ``dimension`` and every value are validated\nagainst the pinned :data:`COGNITIVE_DIMENSIONS` contract."
      },
...
      "ContentConstraint": {
        "properties": {
          "tag_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Type"
          },
          "tag_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Value"
          },
          "tags": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "minimum": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum"
          },
          "maximum": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum"
          },
          "mode": {
            "type": "string",
            "enum": [
              "count",
              "proportion"
            ],
            "title": "Mode",
            "default": "count"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "title": "ContentConstraint",
        "description": "A min/max bound on the items satisfying a tag predicate.\n\nTwo ways to target items (an item must match **all** predicates):\n- **Marginal** (one tag dimension): set ``tag_type`` + ``tag_value`` \u2014 e.g.\n  ``KC=algebra`` or ``Bloom=apply``.\n- **Cross-classified cell** (a content \u00d7 cognitive table cell): set ``tags`` to a\n  mapping of dimension \u2192 value \u2014 e.g. ``{\"KC\": \"algebra\", \"Bloom\": \"apply\"}``\n  means *algebra AND apply*.\n\nBounds are interpreted per ``mode``: ``count`` = absolute item counts;\n``proportion`` = a fraction in [0, 1] of the form length, resolved to a count by\nthe compiler (nearest integer). At least one of ``minimum`` / ``maximum`` set."
      },
...
      "CurriculumManifest": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/ManifestUnit"
            },
            "type": "array",
            "minItems": 1,
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "units"
        ],
        "title": "CurriculumManifest",
        "description": "The minimal curriculum shape the generator reads (derived from unit JSONs)."
      },
...
      "EnemyPolicy": {
        "properties": {
          "enforce": {
            "type": "boolean",
            "title": "Enforce",
            "default": true
          }
        },
        "type": "object",
        "title": "EnemyPolicy",
        "description": "How to honor ``enemy_of`` relations from the bank.\n\nWhen ``enforce`` is true, two items that are enemies of each other may not both\nappear in the same form (declared one-directionally in the bank; the compiler\nsymmetrizes)."
      },
...
      "ExposureFeedback": {
        "properties": {
          "count_contexts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Count Contexts"
          },
          "max_cumulative": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cumulative"
          },
          "prefer_underused": {
            "type": "boolean",
            "title": "Prefer Underused",
            "default": false
          },
          "underuse_weight": {
            "type": "number",
            "minimum": 0.0,
            "title": "Underuse Weight",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "ExposureFeedback",
        "description": "**Opt-in, default-off** feedback from *longitudinal* item-exposure history into\nassembly eligibility.\n\nDistinct from :class:`ExposureTarget` (within-batch overlap/use/rate, governing a\nsingle multi-form assembly) and from CAT administration-time exposure control\n(Sympson-Hetter etc.). Uses cumulative usage *across* past assemblies/publications:\n- ``max_cumulative`` \u2014 hard-exclude items already used at least this many times\n  (over-exposure control).\n- ``prefer_underused`` + ``underuse_weight`` \u2014 bias selection toward under-utilized\n  items (bidirectional utilization). ``underuse_weight`` is in objective info-units\n  per unit of cumulative exposure: small = tie-breaker, large = strong preference.\n\nWhen this is absent the assembly is **byte-for-byte unchanged** (no eligibility\nconstraints/terms are added)."
      },
      "ExposureTarget": {
        "properties": {
          "max_use_per_item": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Use Per Item"
          },
          "max_exposure_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Exposure Rate"
          },
          "max_pairwise_overlap": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pairwise Overlap"
          }
        },
        "type": "object",
        "title": "ExposureTarget",
        "description": "Optional caps on item reuse across assembled forms (multi-form jobs).\n\nTwo reuse levers:\n- **per-item use cap**: an item appears in at most ``max_use_per_item`` forms.\n  Specify it directly, or as a target ``max_exposure_rate`` (proportion 0\u20131) that\n  the compiler translates to a count given the planned form count:\n  ``max_use \u2248 ceil(rate \u00d7 num_forms)`` (assumes **uniform form administration**).\n  A raw ``max_use_per_item`` is the low-level override and wins if both are set.\n- **pairwise overlap cap**: any two forms may share at most\n  ``max_pairwise_overlap`` items (distinct from the total per-item cap \u2014 it bounds\n  similarity *between forms*, e.g. for security across parallel administrations)."
      },
      "FeasibilityIssue": {
        "properties": {
          "constraint_key": {
            "type": "string",
            "title": "Constraint Key"
          },
          "required": {
            "type": "integer",
            "title": "Required"
          },
          "available": {
            "type": "integer",
            "title": "Available"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "constraint_key",
          "required",
          "available",
          "message"
        ],
        "title": "FeasibilityIssue"
      },
...
      "GenerateBlueprintRequest": {
        "properties": {
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurriculumManifest"
              },
              {
                "type": "null"
              }
            ]
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "test_type": {
            "type": "string",
            "enum": [
              "unit_quiz",
              "mid_course",
              "end_of_course",
              "cumulative_final"
            ],
            "title": "Test Type",
            "default": "cumulative_final"
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "scope_unit_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Unit Ids"
          },
          "length": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Length"
          },
          "num_forms": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Num Forms",
            "default": 1
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "binding": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fixed_form",
                  "loft",
                  "cat"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Binding"
          },
          "max_per_complicator": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Max Per Complicator",
            "default": 2
          },
          "constraint_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "count",
                  "proportion"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Mode"
          },
          "statistical_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TIFTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "unit_tag": {
            "type": "string",
            "title": "Unit Tag",
            "default": "unit"
          },
          "kc_tag": {
            "type": "string",
            "title": "Kc Tag",
            "default": "kc"
          },
          "complicator_tag": {
            "type": "string",
            "title": "Complicator Tag",
            "default": "complicator"
          },
          "cognitive_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CognitiveProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "pool_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pool Id"
          }
        },
        "type": "object",
        "required": [
          "length"
        ],
        "title": "GenerateBlueprintRequest",
        "description": "Inputs for one generated blueprint (\u00a76.2: one generator, one weight\nfunction, four scopes).\n\nThe curriculum comes either inline (``manifest``) or from the server catalog\n(``course_id``, see ``GET /curricula``) \u2014 exactly one. ``test_type`` picks the\n\u00a76.2 shape:\n\n- ``unit_quiz`` \u2014 one unit (``unit_id``): per-KC shares \u221d w(KC), plus a\n  per-complicator **maximum** so a form cannot drill one complicator.\n  Default binding: LOFT.\n- ``mid_course`` \u2014 first-half units; ``end_of_course`` \u2014 second-half units;\n  ``cumulative_final`` \u2014 all units: per-unit shares \u221d w(unit), renormalized\n  within scope. Default binding: CAT. ``scope_unit_ids`` overrides the\n  derived scope for any of the three.\n\nWeights follow \u00a76.1 (dimension sums with median imputation). Content-only by\ndefault; ``statistical_target`` optionally attaches a TIF template for\nfixed-form/LOFT bindings (LOFT requires a tolerance, \u00a74.1)."
      },
      "GenerateBlueprintResponse": {
        "properties": {
          "blueprint": {
            "$ref": "#/components/schemas/Blueprint"
          },
          "shares": {
            "items": {
              "$ref": "#/components/schemas/ShareLine"
            },
            "type": "array",
            "title": "Shares"
          },
          "imputed_fraction": {
            "type": "number",
            "title": "Imputed Fraction",
            "default": 0.0
          },
          "feasibility_checked": {
            "type": "boolean",
            "title": "Feasibility Checked"
          },
          "feasible": {
            "type": "boolean",
            "title": "Feasible"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/FeasibilityIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "blueprint",
          "shares",
          "feasibility_checked",
          "feasible",
          "issues",
          "warnings"
        ],
        "title": "GenerateBlueprintResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManifestKC": {
        "properties": {
          "kc_id": {
            "type": "string",
            "title": "Kc Id"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "complicators": {
            "items": {
              "$ref": "#/components/schemas/ManifestComplicator"
            },
            "type": "array",
            "title": "Complicators"
          },
          "n_complicators": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "N Complicators"
          }
        },
        "type": "object",
        "required": [
          "kc_id"
        ],
        "title": "ManifestKC"
      },
      "ManifestUnit": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "kcs": {
            "items": {
              "$ref": "#/components/schemas/ManifestKC"
            },
            "type": "array",
            "minItems": 1,
            "title": "Kcs"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "kcs"
        ],
        "title": "ManifestUnit"
      },
...
      "PerUnitCognitiveMinimum": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "minimum": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Minimum"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "value",
          "minimum"
        ],
        "title": "PerUnitCognitiveMinimum",
        "description": "A cross-classified cell minimum: at least ``minimum`` items in ``unit_id``\ncarrying ``value`` on the profile's dimension."
      },
...
      "ShareLine": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "share": {
            "type": "number",
            "title": "Share"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "n_imputed": {
            "type": "integer",
            "title": "N Imputed",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "key",
          "weight",
          "share",
          "count"
        ],
        "title": "ShareLine",
        "description": "How one unit/KC's item share was derived (weight \u2192 share \u2192 count).\n\n``weight`` is the \u00a76.1 dimension sum (float: the domain-median imputation for\nunknown counts can be fractional). ``n_imputed`` says how many of the row's\ncomplicators had their dimension count imputed rather than known."
      },
...
      "TIFTarget": {
        "properties": {
          "theta_points": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "minItems": 1,
            "title": "Theta Points"
          },
          "target_info": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "minItems": 1,
            "title": "Target Info"
          },
          "method": {
            "type": "string",
            "enum": [
              "minimax",
              "maximin"
            ],
            "title": "Method",
            "default": "minimax"
          },
          "tolerance": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerance"
          },
          "weights": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "theta_points",
          "target_info"
        ],
        "title": "TIFTarget",
        "description": "Target test information at a set of theta points.\n\n``method`` selects the assembly objective: ``minimax`` (drive actual TIF to the\ntarget, minimizing the worst-point absolute miss \u2014 the default for parallel\nforms) or ``maximin`` (maximize information at the worst theta point \u2014 there is\n**no target** under maximin, so ``target_info``/``tolerance``/``weights`` are\nignored). ``tolerance`` is an optional absolute band; when set, the compiler adds\nhard ``|actual - target| <= tolerance`` constraints in addition to the objective.\n\n``weights`` (minimax only) give a per-theta multiplier on the deviation term, so\nthe objective minimizes ``max_k w_k\u00b7|TIF_k \u2212 target_k|``. Default all 1.0 (exactly\nthe unweighted minimax). Raise a point's weight to **protect fit there** (e.g. a\ncut score) when the pool forces tradeoffs \u2014 orthogonal to target height: height\nsets the desired curve *shape*; weight sets *where not to compromise*."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}